pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "A multi-agent system for AI-powered applications with specialized agents for memory, tasks, calendar, and web crawling."
authors = [
    {name = "Ishwanku Saini", email = "ishwanku.saini@eversana.com"}
]
dependencies = [
    "fastapi>=0.92.0",
    "uvicorn>=0.20.0",
    "starlette>=0.25.0",
    "httpx>=0.23.3",
    "sseclient-py>=1.8.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.5",
    "requests>=2.28.2",
    "beautifulsoup4>=4.11.2",
    "aiohttp>=3.8.3",
    "pytest>=7.2.1",
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.118.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.5.1",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.1",
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
strict = true

[tool.pytest.ini_options]
addopts = "--cov=mcp --cov-report=term-missing"
testpaths = ["src/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"